victim/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
httpx==0.25.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
